move server/app to server/index
