+++ server/next.js
@@ -1,5 +1,5 @@
-import App from './app'
+import Server from './'
 
 module.exports = (dir, opts) => {
-  return new App(dir, opts)
+  return new Server(dir, opts)
 }
